# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/auth/register": {
      "post": {
        "description": "Creates a user and returns an access token, so the client is signed in immediately.",
        "operationId": "Auth_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed — `message` lists every problem",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "An account with this email already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded — retry after the window resets",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Register a new account",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "description": "Exchanges email and password for a JWT access token.",
        "operationId": "Auth_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed — `message` lists every problem",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid email or password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded — retry after the window resets",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Log in",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "description": "Handy for checking that a token or API key works.",
        "operationId": "Auth_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentUserDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded — retry after the window resets",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": []
          },
          {
            "JWT": []
          }
        ],
        "summary": "Get the authenticated user",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/api-keys": {
      "post": {
        "description": "Returns the full key **once** — only a SHA-256 hash is stored, so it cannot be recovered later. Send it in the `x-api-key` header to call workflow, execution, scheduler and metrics endpoints.",
        "operationId": "ApiKeys_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApiKeyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedApiKeyDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed — `message` lists every problem",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded — retry after the window resets",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Create an API key",
        "tags": [
          "API Keys"
        ]
      },
      "get": {
        "description": "Keys are masked; only the display prefix is returned.",
        "operationId": "ApiKeys_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApiKeyDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded — retry after the window resets",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "List your API keys",
        "tags": [
          "API Keys"
        ]
      }
    },
    "/api/v1/api-keys/{id}": {
      "delete": {
        "description": "The key stops working immediately.",
        "operationId": "ApiKeys_revoke",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "API key id",
            "schema": {
              "example": "cm0x9c2k10001abcd8f7g6h5j",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Key revoked"
          },
          "401": {
            "description": "Missing, invalid or expired credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found, or not owned by the caller",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded — retry after the window resets",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Revoke an API key",
        "tags": [
          "API Keys"
        ]
      }
    },
...
  },
  "info": {
    "title": "CronFlow API",
    "description": "CronFlow runs multi-step workflows — HTTP calls and delays — on a **cron schedule**, from a **webhook**, or **on demand**.\n\n### Quick start\n1. `POST /api/v1/auth/register` and copy the `accessToken`\n2. Click **Authorize**, paste the token into **JWT**\n3. `POST /api/v1/workflows` to create a workflow (see the request examples)\n4. `POST /api/v1/workflows/{workflowId}/executions` to run it now, then poll `GET /api/v1/executions/{id}`\n\n### Authentication\n| Method | Header | Obtain from |\n| --- | --- | --- |\n| JWT | `Authorization: Bearer <accessToken>` | `POST /api/v1/auth/login` |\n| API key | `x-api-key: cf_…` | `POST /api/v1/api-keys` |\n\n### Errors\nEvery error response uses the same envelope (`ErrorResponseDto`). Its `requestId` matches the\n`X-Request-Id` response header and the server logs.\n\n### Rate limiting\nRequests are limited per client IP (100/minute by default; 10/minute for register and login).\nExceeding the limit returns `429 Too Many Requests`.",
    "version": "3.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Access token from POST /api/v1/auth/login"
      },
      "ApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API key from POST /api/v1/api-keys"
      }
    },
    "schemas": {
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "error": {
            "type": "string",
            "example": "Bad Request",
            "description": "HTTP reason phrase"
          },
          "message": {
            "description": "A single message, or one message per failed validation rule",
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ],
            "example": [
              "cronExpression must be a valid 5-field cron expression (minute hour day-of-month month day-of-week)"
            ]
          },
          "path": {
            "type": "string",
            "example": "/api/v1/workflows"
          },
          "method": {
            "type": "string",
            "example": "POST"
          },
          "timestamp": {
            "type": "string",
            "example": "2026-09-14T10:00:00.000Z",
            "format": "date-time"
          },
          "requestId": {
            "type": "string",
            "example": "5f0e7c1a-3b8e-4d7a-9c2f-1e6b8a4d2c10",
            "description": "Correlates this response with server logs (also sent as the X-Request-Id header)"
          }
        },
        "required": [
          "statusCode",
          "error",
          "message",
          "path",
          "method",
          "timestamp"
        ]
      },
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "ada@example.com",
            "maxLength": 254,
            "description": "Stored lower-cased"
          },
          "password": {
            "type": "string",
            "example": "correct-horse-battery-staple",
            "minLength": 8,
            "description": "At least 8 characters and at most 72 bytes"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "AuthUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cm0x8b1f40000abcdlkj2h3g4"
          },
          "email": {
            "type": "string",
            "example": "ada@example.com"
          }
        },
        "required": [
          "id",
          "email"
        ]
      },
      "AuthResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJjbTB4OGIxZjQwMDAwIn0.signature"
          },
          "tokenType": {
            "type": "string",
            "example": "Bearer",
            "enum": [
              "Bearer"
            ]
          },
          "expiresIn": {
            "type": "number",
            "example": 86400,
            "description": "Seconds until the access token expires"
          },
          "user": {
            "$ref": "#/components/schemas/AuthUserDto"
          }
        },
        "required": [
          "accessToken",
          "tokenType",
          "expiresIn",
          "user"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "ada@example.com"
          },
          "password": {
            "type": "string",
            "example": "correct-horse-battery-staple"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "CurrentUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cm0x8b1f40000abcdlkj2h3g4"
          },
          "email": {
            "type": "string",
            "example": "ada@example.com"
          },
          "authMethod": {
            "type": "string",
            "enum": [
              "jwt",
              "api-key"
            ],
            "example": "jwt",
            "description": "How this request was authenticated"
          }
        },
        "required": [
          "id",
          "email",
          "authMethod"
        ]
      },
      "CreateApiKeyDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "GitHub Actions deploy hook",
            "description": "Human-readable label",
            "maxLength": 80
          }
        },
        "required": [
          "name"
        ]
      },
      "CreatedApiKeyDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cm0x9c2k10001abcd8f7g6h5j"
          },
          "name": {
            "type": "string",
            "example": "GitHub Actions deploy hook"
          },
          "keyPrefix": {
            "type": "string",
            "example": "cf_Q2hhbmdl",
            "description": "First characters of the key, for identification only"
          },
          "lastUsedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "example": null
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-09-14T10:00:00.000Z"
          },
          "key": {
            "type": "string",
            "example": "cf_Q2hhbmdlTWUtVGhpcy1Jcy1Bbi1FeGFtcGxlLUtleQ",
            "description": "The full API key. Returned only once — store it securely."
          }
        },
        "required": [
          "id",
          "name",
          "keyPrefix",
          "lastUsedAt",
          "createdAt",
          "key"
        ]
      },
      "ApiKeyDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cm0x9c2k10001abcd8f7g6h5j"
          },
          "name": {
            "type": "string",
            "example": "GitHub Actions deploy hook"
          },
          "keyPrefix": {
            "type": "string",
            "example": "cf_Q2hhbmdl",
            "description": "First characters of the key, for identification only"
          },
          "lastUsedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "example": null
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-09-14T10:00:00.000Z"
          }
        },
        "required": [
          "id",
          "name",
          "keyPrefix",
          "lastUsedAt",
          "createdAt"
        ]
      },
...
    }
  }
}